AI/requirements.txt
pytesseract==0.3.13
opencv-python==4.10.0.84
pillow==10.4.0
pdf2image==1.17.0
numpy==1.26.4
reportlab==4.2.5
jinja2==3.1.4
python-dateutil==2.9.0
pymongo==4.10.1
motor==3.6.0
pydantic==2.9.2
python-dotenv==1.0.1
fastapi==0.115.0
